% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-evaluation.R
\name{big_upset_df}
\alias{big_upset_df}
\title{Plot the -log10p-value distributions for two univariate GWAS against
one another using ggplot.}
\usage{
big_upset_df(effects, thr = 7, quantile = NA, metadata)
}
\arguments{
\item{effects}{fbm created using 'dive_phe2effects' or 'dive_phe2mash'.
Saved under the name "gwas_effects_{suffix}.rds" and can be loaded into
R using the bigstatsr function "big_attach".}

\item{thr}{Numeric. Threshold above which SNP/row is kept for comparisons.}

\item{quantile}{Numeric. Top quantile/percentile to keep for each GWAS for
comparisons.}

\item{metadata}{Metadata created using 'dive_phe2effects' or 'dive_phe2mash'.
Saved under the name "gwas_effects_{suffix}_associated_metadata.csv".}
}
\description{
This function takes GWAS results from the dive_ functions of
snpdiver that create FBM of univariate GWAS effects. It creates a
dataframe from these results suitable for an Upset plot, containing only
the rows/SNPs significant in at least one univariate GWAS at the
-log10p threshold specified.
}
