% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{check_covmat_basics}
\alias{check_covmat_basics}
\title{Basic sanity check for covariance matrices}
\usage{
check_covmat_basics(x)
}
\arguments{
\item{x}{input matrix}
}
\description{
Basic sanity check for covariance matrices
}
