% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{check_positive_definite}
\alias{check_positive_definite}
\title{check matrix for positive definitness}
\usage{
check_positive_definite(x)
}
\arguments{
\item{x}{input matrix}
}
\description{
check matrix for positive definitness
}
