% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_phe2effects.R
\name{dive_phe2effects}
\alias{dive_phe2effects}
\title{Wrapper to run mash given a phenotype data frame}
\usage{
dive_phe2effects(
  df,
  snp,
  type = "linear",
  svd = NULL,
  suffix = "",
  outputdir = ".",
  min.phe = 200,
  save.plots = TRUE,
  thr.r2 = 0.2,
  roll.size = 50,
  verbose = TRUE
)
}
\arguments{
\item{df}{Dataframe containing phenotypes for mash where the first column is
'sample.ID', which should match values in the snp$fam$sample.ID column.}

\item{snp}{A "bigSNP" object; load with \code{snp_attach()}.}

\item{type}{Character string, or a character vector the length of the number
of phenotypes. Type of univarate regression to run for GWAS.
Options are "linear" or "logistic".}

\item{svd}{A "big_SVD" object; Optional covariance matrix to use for
population structure correction.}

\item{suffix}{Optional character vector to give saved files a unique search
string/name.}

\item{outputdir}{Optional file path to save output files.}

\item{min.phe}{Integer. Minimum number of individuals phenotyped in order to
include that phenotype in GWAS. Default is 200. Use lower values with
caution.}

\item{save.plots}{Logical. Should Manhattan and QQ-plots be generated and
saved to the working directory for univariate GWAS? Default is TRUE.}

\item{thr.r2}{Value between 0 and 1. Threshold of r2 measure of linkage
disequilibrium. Markers in higher LD than this will be subset using clumping.}

\item{roll.size}{Integer. Used to create the svd for GWAS.}

\item{verbose}{Output some information on the iterations? Default is \code{TRUE}.}
}
\value{
A mash object made up of all phenotypes where univariate GWAS ran
successfully.
}
\description{
This function allows
you to go from a phenotype data.frame of a few phenotypes you want to
compare to filebacked matrix of univariate GWAS effects, standard errors,
and -log10pvalues. This output object can be used in "dive_effects2mash"
function. Some exception handling has been built into
this function, but the user should stay cautious and skeptical of any
results that seem 'too good to be true'.
}
