% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash-evaluation.R
\name{get_marker_df}
\alias{get_marker_df}
\title{Get mash marker_df}
\usage{
get_marker_df(m, snp)
}
\arguments{
\item{m}{An object of type mash}

\item{snp}{A bigSNP object, produced by the bigsnpr package. Here, the WAMI
SNP information.}
}
\description{
Pulls SNP markers information in the mash object from a bigsnp
object.
}
