% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash-evaluation.R
\name{get_top_results_clumped}
\alias{get_top_results_clumped}
\title{Clump SNPs from a mash object based on -log10BF}
\usage{
get_top_results_clumped(m, snp, thr.r2)
}
\arguments{
\item{m}{A mash object (outputted by mash).}

\item{snp}{A bigSNP object, produced by the bigsnpr package. Here,
the WAMI SNP information.}

\item{thr.r2}{Value between 0 and 1. Threshold of r2 measure of linkage
disequilibrium. Markers in higher LD than this will be subset using clumping.}
}
\description{
Clump SNPs from a mash object based on -log10BF
}
