% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{labelled_stop}
\alias{labelled_stop}
\title{A wrapper function to \code{stop} call}
\usage{
labelled_stop(x, msg)
}
\arguments{
\item{x}{input matrix}

\item{msg}{Character string. A message to append to the stop call.}
}
\description{
A wrapper function to \code{stop} call
}
