% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash-evaluation.R
\name{mash_plot_pairwise_sharing}
\alias{mash_plot_pairwise_sharing}
\title{Create a ggplot of pairwise sharing of mash effects}
\usage{
mash_plot_pairwise_sharing(
  m = NULL,
  effectRDS = NULL,
  corrmatrix = NULL,
  reorder = TRUE,
  thr.r2 = NA,
  saveoutput = FALSE,
  filename = NA,
  suffix = "",
  ...
)
}
\arguments{
\item{m}{An object of type mash}

\item{effectRDS}{An RDS containing a correlation matrix.}

\item{corrmatrix}{A correlation matrix}

\item{reorder}{Logical. Should the columns be reordered by similarity?}

\item{thr.r2}{Numeric. Value between 0 and 1. The r^2 value above which
effects are considered to be linked; in this case, only the SNP with the
more significant log10Bayes Factor value in mash is kept.}

\item{saveoutput}{Logical. Should the output be saved to the path?}

\item{filename}{Character string with an output filename. Optional.}

\item{suffix}{Character. Optional. A unique suffix used to save the files,
instead of the current date & time.}

\item{...}{Other arguments to \code{\link{get_pairwise_sharing}} or
\code{\link{ggcorr}}.}
}
\value{
A list containing a dataframe containing the correlations and a
ggplot2 object containing the correlation plot.
}
\description{
Given a correlation matrix, an RDS with a correlation matrix, or
a mash object, create a ggplot of pairwise sharing of mash effects using
\code{\link{get_pairwise_sharing}} and \code{\link{ggcorr}}.
}
