% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash-evaluation.R
\name{mash_plot_Ulist}
\alias{mash_plot_Ulist}
\title{ggplot of specific covariance matrix patterns}
\usage{
mash_plot_Ulist(
  m,
  range = NA,
  saveoutput = FALSE,
  suffix = "",
  limits = TRUE,
  labels = TRUE
)
}
\arguments{
\item{m}{An object of type mash}

\item{range}{Numeric vector. Which covariance matrices should be plotted?}

\item{saveoutput}{Logical. Should the output be saved to the path?}

\item{suffix}{Character. Optional. A unique suffix used to save the files,
instead of the current date & time.}

\item{limits}{Logical. Should there be plot limits of -1 and 1? Default is TRUE.}

\item{labels}{Logical. Should each pairwise comparison proportion have a label? Default is TRUE.}
}
\value{
A list of dataframes used to make the tile plots and the plots
themselves.
}
\description{
Creates a tile plot using ggplot of the covariance matrices
specified in the mash model.
}
