% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{printf2}
\alias{printf2}
\title{print message function if verbose}
\usage{
printf2(verbose, ...)
}
\arguments{
\item{verbose}{Logical. If TRUE, print progress messages.}

\item{...}{Other arguments to \code{printf()}}
}
\description{
print message function if verbose
}
